Reorganize component filing structure
