+++ src/App.js
@@ -2,8 +2,8 @@ import './App.css';
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 import { useState } from 'react';
 import { io } from 'socket.io-client';
-import Home from './Components/Home';
-import Chat from './Components/Chat';
+import Home from './Components/Home/Home';
+import Chat from './Components/Chat/Chat';
 
 const socket = io.connect('http://localhost:4000');
